+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.0-SNAPSHOT
 
+* refactor package naming to `rocks.coffeenet.plugin.islieb`
 
 ## 0.1.0